Updated ItemAPI
1. ItemAPI now has appropriate documentation.

2. The style checker does not like internal file sections so item name loading has been moved to its own file.

3. Changed the line endings in .editorconfig to CRLF to comply with the style checker's [NewlineAtEndOfFile] expectations.

ItemAPI and ItemNames are filly compliant with the style checker.
