+++ .editorconfig
@@ -8,7 +8,7 @@ root = true
 charset = utf-8
 indent_style = tab
 indent_size = 4
-end_of_line = lf
+end_of_line = crlf
 insert_final_newline = true
 continuation_indent_size = 8
