+++ src/main/java/vg/civcraft/mc/civmodcore/CivModCorePlugin.java
@@ -2,7 +2,7 @@ package vg.civcraft.mc.civmodcore;
 
 import org.bukkit.configuration.serialization.ConfigurationSerialization;
 import org.bukkit.event.HandlerList;
-import vg.civcraft.mc.civmodcore.api.ItemAPI;
+import vg.civcraft.mc.civmodcore.api.ItemNames;
 import vg.civcraft.mc.civmodcore.chatDialog.ChatListener;
 import vg.civcraft.mc.civmodcore.chatDialog.DialogManager;
 import vg.civcraft.mc.civmodcore.dao.ManagedDatasource;
@@ -44,7 +44,7 @@ public final class CivModCorePlugin extends ACivMod {
 			database = null;
 		}
 		// Load APIs
-		ItemAPI.loadItemNames();
+		ItemNames.loadItemNames();
 		new NiceNames().loadNames();
 		new DialogManager();
 		ConfigurationSerialization.registerClass(ManagedDatasource.class);
@@ -62,7 +62,7 @@ public final class CivModCorePlugin extends ACivMod {
 		// De-register listeners
 		HandlerList.unregisterAll(this);
 		// Unload APIs
-		ItemAPI.resetItemNames();
+		ItemNames.resetItemNames();
 		ChunkMetaAPI.saveAll();
 		chunkMetaManager = null;
 		// Disconnect database
